PRLC100-288 - Refactored & moved common address files
