+++ src/main/steps/c100-rebuild/other-person-details/address/select/content.test.ts
@@ -10,7 +10,7 @@ const en = {
   changePostCodeLabel: 'Change postcode',
   errors: {
     selectAddress: {
-      notSelected: 'Select an address',
+      notSelected: 'Select an address from the list',
     },
   },
 };
@@ -20,7 +20,7 @@ const cy = {
   changePostCodeLabel: 'Change postcode - welsh',
   errors: {
     selectAddress: {
-      notSelected: 'Select an address -  welsh',
+      notSelected: 'Select an address from the list -  welsh',
     },
   },
 };
