+++ src/main/steps/c100-rebuild/people/address/address-manual.ts
@@ -1,12 +1,11 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-import { C100RebuildPartyDetails, YesNoEmpty, YesOrNo } from '../../../../../app/case/definition';
-import { TranslationFn } from '../../../../../app/controller/GetController';
-import { FormContent } from '../../../../../app/form/Form';
-import { isFieldFilledIn, isInvalidPostcode } from '../../../../../app/form/validation';
+import { C100RebuildPartyDetails, YesNoEmpty, YesOrNo } from '../../../../app/case/definition';
+import { TranslationFn } from '../../../../app/controller/GetController';
+import { FormContent } from '../../../../app/form/Form';
+import { isFieldFilledIn, isInvalidPostcode } from '../../../../app/form/validation';
 
 const en = () => ({
   addressLine1: 'Building and street',
-  addressLine1Hint: 'Court documents may be sent here',
   town: 'Town or city',
   county: 'County',
   country: 'Country',
@@ -17,7 +16,6 @@ const en = () => ({
 
 const cy = () => ({
   addressLine1: 'Building and street - welsh',
-  addressLine1Hint: 'Court documents may be sent here - welsh',
   town: 'Town or city - welsh',
   county: 'County - welsh',
   country: 'Country - welsh',
