+++ src/main/steps/c100-rebuild/people/address/address-select.ts
@@ -1,7 +1,7 @@
-import { C100RebuildPartyDetails } from '../../../../../app/case/definition';
-import { TranslationFn } from '../../../../../app/controller/GetController';
-import { FormContent } from '../../../../../app/form/Form';
-import { isAddressSelected } from '../../../../../app/form/validation';
+import { C100RebuildPartyDetails } from '../../../../app/case/definition';
+import { TranslationFn } from '../../../../app/controller/GetController';
+import { FormContent } from '../../../../app/form/Form';
+import { isAddressSelected } from '../../../../app/form/validation';
 
 const getAddressItems = addresses => addresses.map((item, index) => ({ text: item.fullAddress, value: index }));
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
